pyproject.toml
[build-system]
requires = [
  "setuptools>=70.0.0", # CVE-2024-38335 recommends this
  "mpmath",
  "numpy",
  "requests",
  "sympy",
  "pygments",
  "mathics-pygments",
  "trepan3k",
  "Mathics3-Module-Base",
  # Until the 9.0 release...
  "Mathics_Scanner @ git+https://github.com/Mathics3/mathics-scanner.git@master",
  "Mathics3 @ git+https://github.com/Mathics3/mathics-core.git@master"
]
build-backend = "setuptools.build_meta"

[project]
name = "Mathics3-trepan"
description = 'Mathics3 Debugger Module based on trepan debuggers'
dependencies = [
        "Mathics_Scanner",
        "Mathics3",
        "Mathics3-Module-Base",
        "mathics-pygments",
        "trepan3k>=1.4.0",
]
requires-python = ">=3.9"
readme = "README.rst"
license = "GPL-3.0-or-later"
keywords = ["Mathematica", "Wolfram", "Interpreter", "Shell", "Math", "CAS"]
maintainers = [
    {name = "Mathics Group", email = "mathics-devel@googlegroups.com"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Interpreters",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/Mathics3/Mathics3-trepan"
Downloads = "https://github.com/Mathics3/Mathics3-trepan/releases"

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
]

[tool.setuptools]
packages = [
    "pymathics.trepan",
]

[tool.setuptools.dynamic]
version = {attr = "pymathics.trepan.__version__"}
